requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
absl-py=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.12.0=pypi_0
argcomplete=3.6.3=pypi_0
asttokens=3.0.1=pypi_0
attrs=25.4.0=pypi_0
blas=1.0=mkl
blinker=1.9.0=pypi_0
bracex=2.6=pypi_0
brotlicffi=1.2.0.0=py310h885b0b7_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
certifi=2025.11.12=py310haa95532_0
cffi=2.0.0=py310h02ab6af_1
charset-normalizer=3.4.4=py310haa95532_0
click=8.3.1=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.10.1=pypi_0
comm=0.2.3=pypi_0
configargparse=1.7.1=pypi_0
contourpy=1.3.2=pypi_0
cuda-cccl=12.9.27=0
cuda-cccl_win-64=12.9.27=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.9.79=0
cuda-runtime=11.8.0=0
cuda-version=12.9=3
cycler=0.12.1=pypi_0
dash=3.4.0=pypi_0
decorator=5.2.1=pypi_0
docstring-parser=0.17.0=pypi_0
einops=0.8.1=pypi_0
evo=1.34.2=pypi_0
exceptiongroup=1.3.1=pypi_0
executing=2.2.1=pypi_0
expat=2.7.3=h885b0b7_4
fastapi=0.128.0=pypi_0
fastjsonschema=2.21.2=pypi_0
ffmpy=1.0.0=pypi_0
filelock=3.20.0=py310haa95532_0
flask=3.1.2=pypi_0
flatbuffers=25.12.19=pypi_0
fonttools=4.61.1=pypi_0
freetype=2.13.3=h0620614_0
fsspec=2025.12.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.2=py310h8598115_0
gradio=5.17.1=pypi_0
gradio-client=1.7.1=pypi_0
grpcio=1.76.0=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.2.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.2.3=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.11=py310haa95532_0
imageio=2.37.2=pypi_0
importlib-metadata=8.7.1=pypi_0
intel-openmp=2021.4.0=pypi_0
iopath=0.1.10=pypi_0
ipython=8.38.0=pypi_0
ipywidgets=8.1.8=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=py310haa95532_0
joblib=1.5.3=pypi_0
jpeg=9f=ha349fce_0
jsonschema=4.26.0=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter-core=5.9.1=pypi_0
jupyterlab-widgets=3.0.16=pypi_0
kiwisolver=1.4.9=pypi_0
kornia=0.8.2=pypi_0
kornia-rs=0.1.10=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.16=hb4a4139_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.10.19=0
libcurand-dev=10.3.10.19=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libexpat=2.7.3=h885b0b7_4
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.50=h46444df_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp-base=1.6.0=hbf3958f_0
libzlib=1.3.1=h02ab6af_0
lightglue=0.0=pypi_0
lxml=6.0.2=pypi_0
lz4=4.4.5=pypi_0
lz4-c=1.9.4=h2bbff1b_1
manifold3d=3.3.2=pypi_0
mapbox-earcut=2.0.0=pypi_0
markdown=3.10.1=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.8=pypi_0
matplotlib-inline=0.2.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py310haa95532_0
msgspec=0.20.0=pypi_0
narwhals=2.15.0=pypi_0
natsort=8.4.0=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=py310haa95532_0
ninja=1.13.0=pypi_0
nodeenv=1.10.0=pypi_0
numexpr=2.14.1=pypi_0
numpy=1.26.1=pypi_0
nvidia-ml-py=13.590.48=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.23.2=pypi_0
open3d=0.19.0=pypi_0
opencv-python=4.8.0.76=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.18=h543e019_0
orjson=3.11.6=pypi_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
parso=0.8.5=pypi_0
pillow=11.1.0=py310h096bfcc_0
pip=25.3=pyhc872135_0
platformdirs=4.5.1=pypi_0
plotly=6.5.2=pypi_0
plyfile=1.1.3=pypi_0
portalocker=3.2.0=pypi_0
prompt-toolkit=3.0.52=pypi_0
protobuf=6.33.5=pypi_0
psutil=7.2.1=pypi_0
pure-eval=0.2.3=pypi_0
pyceres=2.3=pypi_0
pycollada=0.9.3=pypi_0
pycolmap=3.10.0=pypi_0
pycparser=2.23=py310haa95532_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.2=pypi_0
pynvml=13.0.1=pypi_0
pyparsing=3.3.2=pypi_0
pyreadline3=3.5.4=pypi_0
pysocks=1.7.1=py310haa95532_1
python=3.10.19=h981015d_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.22=pypi_0
pytorch=2.3.1=py3.10_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pywin32=311=pypi_0
pyyaml=6.0.3=py310hb9a58be_0
referencing=0.37.0=pypi_0
requests=2.32.5=py310haa95532_1
retrying=1.4.2=pypi_0
rich=14.3.1=pypi_0
rosbags=0.11.0=pypi_0
rpds-py=0.30.0=pypi_0
rtree=1.4.1=pypi_0
ruamel-yaml=0.19.1=pypi_0
ruff=0.14.14=pypi_0
safehttpx=0.1.7=pypi_0
safetensors=0.7.0=pypi_0
scikit-image=0.25.2=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=80.9.0=py310haa95532_0
shapely=2.1.2=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=py310haa95532_0
sqlite=3.51.0=hda9a48d_0
stack-data=0.6.3=pypi_0
starlette=0.50.0=pypi_0
svg-path=7.0=pypi_0
sympy=1.14.0=py310haa95532_1
tabulate=0.9.0=pypi_0
tbb=2021.13.1=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=3.3.0=pypi_0
tifffile=2025.5.10=pypi_0
tk=8.6.15=hf199647_0
tomlkit=0.13.3=pypi_0
torchaudio=2.3.1=pypi_0
torchvision=0.18.1=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
trimesh=4.11.1=pypi_0
typeguard=4.4.4=pypi_0
typer=0.21.1=pypi_0
typer-slim=0.20.0=pypi_0
typing_extensions=4.15.0=py310haa95532_0
tyro=1.0.5=pypi_0
tzdata=2025.3=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.6.1=py310haa95532_0
uvicorn=0.40.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vhacdx=0.0.10=pypi_0
viser=0.2.23=pypi_0
vs2015_runtime=14.44.35208=ha6b5a95_10
wcmatch=10.1=pypi_0
wcwidth=0.5.0=pypi_0
websockets=14.2=pypi_0
werkzeug=3.1.5=pypi_0
wheel=0.45.1=py310haa95532_0
widgetsnbextension=4.0.15=pypi_0
win_inet_pton=1.1.0=py310haa95532_1
xxhash=3.6.0=pypi_0
xz=5.6.4=h4754444_1
yacs=0.1.8=pypi_0
yaml=0.2.5=he774522_0
yourdfpy=0.0.60=pypi_0
zipp=3.23.0=pypi_0
zlib=1.3.1=h02ab6af_0
zstandard=0.25.0=pypi_0
zstd=1.5.7=h56299aa_0
